fixed warning for fast_bilateral_filter
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@7007 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ filters/include/pcl/filters/impl/fast_bilateral.hpp
@@ -80,30 +80,30 @@ pcl::FastBilateralFilter<PointT>::applyFilter (PointCloud &output)
   const size_t padding_xy = 2;
   const size_t padding_z  = 2;
 
-  const size_t small_width  = static_cast<size_t> ((input_->width  - 1) / sigma_s_) + 1 + 2 * padding_xy;
-  const size_t small_height = static_cast<size_t> ((input_->height - 1) / sigma_s_) + 1 + 2 * padding_xy;
+  const size_t small_width  = static_cast<size_t> (static_cast<float> (input_->width  - 1) / sigma_s_) + 1 + 2 * padding_xy;
+  const size_t small_height = static_cast<size_t> (static_cast<float> (input_->height - 1) / sigma_s_) + 1 + 2 * padding_xy;
   const size_t small_depth  = static_cast<size_t> (base_delta / sigma_r_)   + 1 + 2 * padding_z;
 
 
   Array3D data (small_width, small_height, small_depth);
   for (size_t x = 0; x < input_->width; ++x)
   {
-    const size_t small_x = static_cast<size_t> (x / sigma_s_ + 0.5) + padding_xy;
+    const size_t small_x = static_cast<size_t> (static_cast<float> (x) / sigma_s_ + 0.5f) + padding_xy;
     for (size_t y = 0; y < input_->height; ++y)
     {
       const float z = output (x,y).z - base_min;
 
-      const size_t small_y = static_cast<size_t> (y / sigma_s_ + 0.5) + padding_xy;
-      const size_t small_z = static_cast<size_t> (z / sigma_r_ + 0.5) + padding_z;
+      const size_t small_y = static_cast<size_t> (static_cast<float> (y) / sigma_s_ + 0.5f) + padding_xy;
+      const size_t small_z = static_cast<size_t> (static_cast<float> (z) / sigma_r_ + 0.5f) + padding_z;
 
       Eigen::Vector2f& d = data (small_x, small_y, small_z);
       d[0] += output (x,y).z;
-      d[1] += 1.0;
+      d[1] += 1.0f;
     }
   }
 
 
-  std::vector<int> offset (3);
+  std::vector<long int> offset (3);
   offset[0] = &(data (1,0,0)) - &(data (0,0,0));
   offset[1] = &(data (0,1,0)) - &(data (0,0,0));
   offset[2] = &(data (0,0,1)) - &(data (0,0,0));
@@ -112,7 +112,7 @@ pcl::FastBilateralFilter<PointT>::applyFilter (PointCloud &output)
 
   for (size_t dim = 0; dim < 3; ++dim)
   {
-    const int off = offset[dim];
+    const long int off = offset[dim];
     for (size_t n_iter = 0; n_iter < 2; ++n_iter)
     {
       std::swap (buffer, data);
@@ -194,19 +194,19 @@ pcl::FastBilateralFilter<PointT>::Array3D::trilinear_interpolation (const float 
   const size_t z_index  = clamp (0, z_dim_ - 1, static_cast<size_t> (z));
   const size_t zz_index = clamp (0, z_dim_ - 1, z_index + 1);
 
-  const float x_alpha = x - x_index;
-  const float y_alpha = y - y_index;
-  const float z_alpha = z - z_index;
+  const float x_alpha = x - static_cast<float> (x_index);
+  const float y_alpha = y - static_cast<float> (y_index);
+  const float z_alpha = z - static_cast<float> (z_index);
 
   return
       (1.0f-x_alpha) * (1.0f-y_alpha) * (1.0f-z_alpha) * (*this)(x_index, y_index, z_index) +
-      x_alpha        * (1.0f-y_alpha) * (1.0f-z_alpha) * (*this)(xx_index,y_index, z_index) +
-      (1.0f-x_alpha) * y_alpha        * (1.0f-z_alpha) * (*this)(x_index, yy_index,z_index) +
-      x_alpha        * y_alpha        * (1.0f-z_alpha) * (*this)(xx_index,yy_index,z_index) +
+      x_alpha        * (1.0f-y_alpha) * (1.0f-z_alpha) * (*this)(xx_index, y_index, z_index) +
+      (1.0f-x_alpha) * y_alpha        * (1.0f-z_alpha) * (*this)(x_index, yy_index, z_index) +
+      x_alpha        * y_alpha        * (1.0f-z_alpha) * (*this)(xx_index, yy_index, z_index) +
       (1.0f-x_alpha) * (1.0f-y_alpha) * z_alpha        * (*this)(x_index, y_index, zz_index) +
-      x_alpha        * (1.0f-y_alpha) * z_alpha        * (*this)(xx_index,y_index, zz_index) +
-      (1.0f-x_alpha) * y_alpha        * z_alpha        * (*this)(x_index, yy_index,zz_index) +
-      x_alpha        * y_alpha        * z_alpha        * (*this)(xx_index,yy_index,zz_index);
+      x_alpha        * (1.0f-y_alpha) * z_alpha        * (*this)(xx_index, y_index, zz_index) +
+      (1.0f-x_alpha) * y_alpha        * z_alpha        * (*this)(x_index, yy_index, zz_index) +
+      x_alpha        * y_alpha        * z_alpha        * (*this)(xx_index, yy_index, zz_index);
 }
 
 #endif /* FAST_BILATERAL_HPP_ */
